Configure GitHub Pages deployment and update app for static export
