+++ noaa-marine-weather/.gitignore
@@ -15,7 +15,8 @@
 
 # next.js
 /.next/
-/out/
+# Don't ignore /out/ for GitHub Pages deployment
+# /out/
 
 # production
 /build
@@ -28,10 +29,10 @@
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
-.pnpm-debug.log*
 
-# env files (can opt-in for committing if needed)
-.env*
+# local env files
+.env*.local
+.env
 
 # vercel
 .vercel
