+++ noaa-marine-weather/package.json
@@ -3,28 +3,29 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "dev": "next dev --turbopack",
+    "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "lint": "next lint"
+    "lint": "next lint",
+    "export": "next build"
   },
   "dependencies": {
-    "chart.js": "^4.5.0",
-    "date-fns": "^4.1.0",
+    "react": "19.0.0",
+    "react-dom": "19.0.0",
     "next": "15.4.4",
-    "react": "19.1.0",
-    "react-chartjs-2": "^5.3.0",
-    "react-dom": "19.1.0"
+    "chart.js": "^4.4.6",
+    "react-chartjs-2": "^5.2.0",
+    "date-fns": "^4.1.0"
   },
   "devDependencies": {
-    "@eslint/eslintrc": "^3",
-    "@tailwindcss/postcss": "^4",
+    "typescript": "^5",
     "@types/node": "^20",
     "@types/react": "^19",
     "@types/react-dom": "^19",
+    "@tailwindcss/postcss": "^4.2.4",
+    "tailwindcss": "^4.2.4",
     "eslint": "^9",
     "eslint-config-next": "15.4.4",
-    "tailwindcss": "^4",
-    "typescript": "^5"
+    "@eslint/eslintrc": "^3"
   }
 }
